+++ server/avista_portal/api/__init__.py
@@ -1,4 +1,6 @@
 """ This package contains the modules containing the route logic for the service """
 from flask import Blueprint
 
-api_bp = Blueprint('api', __name__)
+api_bp = Blueprint('portal_api', __name__)
+
+from avista_portal.api import data