+++ server/tests/test_server.py
@@ -2,8 +2,7 @@ import unittest
 
 
 class ServerTest(unittest.TestCase):
-    def test_something(self):
-        self.assertEqual(True, False)
+    pass
 
 
 if __name__ == '__main__':