+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* Added the basic data route to allow for data to enter the portal from the IoT devices
 
 ### Changed
+* Updated existing tests
+* Modified the existing server to correspond to changes in base-server
 
 ### Deleted